social_media_api/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.15
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
python-dotenv>=1.0
psycopg2-binary>=2.9
